One payment methods call for all express payments (#1211)
* chore: comment invoking payment methods in cart
* chore: adapt all payment methods to use the payments method response
