+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/__tests__/applePayExpress.test.js
@@ -8,8 +8,7 @@ const {
   handleApplePayResponse,
   callPaymentFromComponent,
   formatCustomerObject,
-} = require('../../applePayExpressCommon');
-
+} = require('../../applePayExpress');
 
 beforeEach(() => {
   jest.clearAllMocks();
