+++ src/cartridges/app_adyen_SFRA/cartridge/templates/default/cart/checkoutButtons.isml
@@ -86,17 +86,7 @@
                 <div data-method="amazonpay" class="expressComponent" id="amazonpay-container"></div>
                 <div data-method="paypal" class="expressComponent" id="paypal-container" style="padding:0"></div>
             </div>
-            <script src="${URLUtils.httpsStatic('/js/expressPaymentMethodsVisibility.js')}" type="text/javascript"></script>
-        </isif>
-        <script src="${URLUtils.httpsStatic('/js/main.js')}" type="text/javascript"></script>
-        <isif condition="${AdyenConfigs.areExpressPaymentsEnabled() && AdyenConfigs.isApplePayExpressEnabled()}">
-            <script src="${URLUtils.httpsStatic('/js/applePayExpress.js')}" type="text/javascript"></script>
-        </isif>
-        <isif condition="${AdyenConfigs.areExpressPaymentsEnabled() && AdyenConfigs.isAmazonPayExpressEnabled()}">
-            <script src="${URLUtils.httpsStatic('/js/amazonPayExpressPart1.js')}" type="text/javascript"></script>
-        </isif>
-        <isif condition="${AdyenConfigs.areExpressPaymentsEnabled() && AdyenConfigs.isPayPalExpressEnabled()}">
-            <script src="${URLUtils.httpsStatic('/js/paypalExpress.js')}" type="text/javascript"></script>
+            <script src="${URLUtils.httpsStatic('/js/expressPayments.js')}" type="text/javascript"></script>
         </isif>
     </iselse>
 </isif>
